If file resolution fails, try falling back to node's own resolution
This is useful because the implementation here is limited and doesn't
support various new node features, such as export maps. Fixing this is a
very substantial amount of work, that would be complex given the
additional non-standard features and configuration that this module
_does_ support.

Instead, we just let the existing code take priority, and use this only
for cases that are otherwise unresolvable.